+++ Plutarch/Internal/Term.hs
@@ -138,6 +138,8 @@ data RawTerm
   | RPlaceHolder Integer
   | RConstr Word64 [RawTerm]
   | RCase RawTerm [RawTerm]
+  | -- Let x (\x' -> ...)
+    RLet RawTerm RawTerm
   deriving stock (Show, Eq)
 
 -- | A very cheap hash which cheapens equality, but is also needed for using an unordered container.
@@ -158,12 +160,14 @@ instance Hashable RawTerm where
     RPlaceHolder x -> hash (10 :: Int, x)
     RConstr x y -> hash (11 :: Int, x, y)
     RCase x y -> hash (12 :: Int, x, y)
+    RLet x y -> hash (13 :: Int, x, y)
   {-# INLINE hash #-}
 
 data TermResult = TermResult
   { getTerm :: RawTerm
   , getDeps :: [HoistedTerm]
   }
+  deriving stock (Show)
 
 mapTerm :: (RawTerm -> RawTerm) -> TermResult -> TermResult
 mapTerm f (TermResult t d) = TermResult (f t) d
@@ -640,16 +644,24 @@ plam' f = Term $ do
 plet :: Term s a -> (Term s a -> Term s b) -> Term s b
 plet v f = Term $ do
   env <- ask
-  let res = runReaderT (asRawTerm v) env
-  case res of
-    Left _ -> lift res
-    Right (TermResult t _) -> case t of
-      -- Inline sufficiently small terms in WHNF
-      RVar _ -> asRawTerm (f v)
-      RBuiltin _ -> asRawTerm (f v)
-      RHoisted _ -> asRawTerm (f v)
-      -- Do the lambda transform as normal
-      _ -> asRawTerm (papp (plam' f) v)
+  let vRes = runReaderT (asRawTerm v) env
+  case vRes of
+    Left msg -> lift . Left $ "plet failed: " <> msg
+    Right (TermResult vt _vDeps) -> do
+      let doInline = asRawTerm (f v)
+          noInline (TermResult ft _fDeps) = case ft of
+            RError -> pure $ mkTermRes RError
+            RLamAbs 0 (RVar 0) -> asRawTerm v
+            RHoisted (HoistedTerm _ (RLamAbs 0 (RVar 0))) -> asRawTerm v
+            RApply xl xr -> pure . TermResult (RApply xl (vt : xr)) $ _fDeps <> _vDeps
+            _ -> pure $ TermResult (RLet vt ft) (_fDeps <> _vDeps)
+      case vt of
+        -- Inline sufficiently small terms in WHNF
+        RVar _ -> doInline
+        RBuiltin _ -> doInline
+        RHoisted _ -> doInline
+        RError -> pure . mkTermRes $ RError
+        _ -> noInline =<< asRawTerm (plam' f)
 
 pthrow :: HasCallStack => Text -> Term s a
 pthrow msg = Term . lift . Left $ fromString (prettyCallStack callStack) <> "\n\n" <> msg
@@ -833,6 +845,8 @@ rawTermToUPLC m l (RConstr i xs) = UPLC.Constr () i (rawTermToUPLC m l <$> xs)
 rawTermToUPLC m l (RCase x xs) = UPLC.Case () (rawTermToUPLC m l x) $ V.fromList (rawTermToUPLC m l <$> xs)
 -- rawTermToUPLC m l (RHoisted hoisted) = UPLC.Var () . DeBruijn . Index $ l - m hoisted
 rawTermToUPLC m l (RHoisted hoisted) = m hoisted l -- UPLC.Var () . DeBruijn . Index $ l - m hoisted
+-- The second part of a let bind is the "function part" (to follow ordinary `let` semantics from Haskell & etc)
+rawTermToUPLC m l (RLet v f) = rawTermToUPLC m l (RApply f [v])
 
 smallEnoughToInline :: RawTerm -> Bool
 smallEnoughToInline = \case