+++ Plutarch/TermCont.hs
@@ -7,7 +7,6 @@ module Plutarch.TermCont (
   TC.runTermCont,
   TC.unTermCont,
   TC.tcont,
-  TC.pfindPlaceholder,
   pletC,
   pmatchC,
   ptraceC,