+++ plutarch-testlib/test/Test.hs
@@ -34,6 +34,7 @@ import Plutarch.Test.Suite.PlutarchLedgerApi.V1 qualified as V1
 import Plutarch.Test.Suite.PlutarchLedgerApi.V2 qualified as V2
 import Plutarch.Test.Suite.PlutarchLedgerApi.V3 qualified as V3
 import Plutarch.Test.Suite.PlutarchLedgerApi.V3.Contexts qualified as V3.Contexts
+import Plutarch.Test.Suite.PlutarchLedgerApi.V3.Value qualified as V3.Value
 import Plutarch.Test.Suite.PlutarchLedgerApi.Value.AssetClass qualified as AssetClass
 import Test.Tasty (adjustOption, defaultMain, testGroup)
 import Test.Tasty.QuickCheck (QuickCheckTests)
@@ -85,6 +86,7 @@ main = do
               , V2.tests
               , V3.tests
               , PLAParse.tests
+              , V3.Value.tests
               ]
         , adjustOption moreTests $
             testGroup "Regressions" Regressions.tests